requirements.txt
absl-py==0.2.0
amqp==2.3.2
appnope==0.1.0
asn1crypto==0.24.0
astor==0.6.2
backcall==0.1.0
beautifulsoup4==4.6.0
billiard==3.5.0.4
bleach==1.5.0
bottle==0.12.13
bs4==0.0.1
cachetools==2.1.0
certifi==2018.4.16
cffi==1.11.5
chainer==4.2.0
chardet==3.0.4
click==6.7
cryptography==2.3.1
cycler==0.10.0
Cython==0.28.4
decorator==4.3.0
dill==0.2.8.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.1
django-bootstrap4==0.0.6
django-cors-headers==2.4.0
django-filters==0.2.1
django-heroku==0.3.1
django-toolbelt==0.0.1
djangorestframework==3.8.2
filelock==3.0.4
gast==0.2.0
grpcio==1.14.2
gunicorn==19.9.0
html5lib==0.9999999
idna==2.7
kiwisolver==1.0.1
kombu==4.2.1
Markdown==2.6.11
matplotlib==3.0.1
numpy==1.14.3
opencv-python==3.4.2.17
Pillow==5.2.0
protobuf==3.6.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycparser==2.18
PyMySQL==0.9.2
pyparsing==2.2.2
python-dateutil==2.7.4
pytz==2018.5
redis==2.10.6
rq==0.12.0
scikit-learn==0.19.2
scipy==1.1.0
six==1.11.0
sklearn==0.0
static3==0.7.0
tensorboard==1.10.0
tensorflow==1.10.1
termcolor==1.1.0
vine==1.1.4
Werkzeug==0.14.1
whitenoise==4.0
